refactor: 디렉토리 구조 변경
handler 디렉토리 추가
